Add disable_ssl and endpoint config options
Both of these affect the AWS session configuration and can be used
for local testing (i.e. localstack).

If disable_ssl is set to true, SSL will not be used to communicate
with the AWS API.

If endpoint is set it will configure the AWS SDK to communicate with
an alternate AWS API endpoint.